Improve test quality and fix BulkResult.ModifiedCount type
- Replace abbreviated variable names (emp1/emp2) with descriptive names
  (employeeAlice/employeeBob/employeeTarget)
- Batch separate AddAsync calls into single AddAsync(List<T>) calls
- Change BulkResult.ModifiedCount from int to long for consistency with
  PatchAsync(Ids) and PatchAllAsync return types

+++ src/Foundatio.Repositories.Elasticsearch/Repositories/BulkResult.cs
@@ -18,7 +18,7 @@ internal sealed record BulkResult
     public IReadOnlySet<string> RetryableIds { get; init; } = _emptySet;
     public IReadOnlySet<string> FatalIds { get; init; } = _emptySet;
 
-    public int ModifiedCount => SuccessfulIds.Count - NoopIds.Count;
+    public long ModifiedCount => SuccessfulIds.Count - NoopIds.Count;
 
     public string TransportError { get; init; }
     public Exception TransportException { get; init; }
